fix(ci): split admin shadow daily test
